Rename material files
